requirements-metal.txt
# AI & Acceleration (Apple Silicon)
torch==2.2.2
transformers>=4.40.0
onnxruntime>=1.17.0
mlx-lm>=0.1.0

# Data Science
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
matplotlib>=3.7.0

# API & Server
litestar>=2.8.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.29
pgvector>=0.2.5
alembic>=1.13.0

# Trading & Finance
alpaca-py>=0.23.0
yfinance>=0.2.37

# Telemetry (OpenTelemetry)
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-asyncpg

# Agent Framework
langgraph>=0.0.10
langchain>=0.1.0
langchain-core>=0.1.0
google-generativeai>=0.5.0
